refactored to model.py
