+++ examples/boltzmann_wealth_model/wealth_model/model.py
